vaev+karm: Update from upstream.
meta: Removed karm from the repository.

meta: Removed useless compile_flags.txt

Please use `./ck b b --database` instead

meta: Moved Karm, Hideo, and Vaev to separate repositories.

meta: Update to karm 0.1.0.

meta: Update externs.

meta: Removed useless readme.md

meta: Bring all ma boi home.

+++ src/libs/karm-cli/args.cpp
@@ -40,11 +40,11 @@ Res<bool> ValueParser<bool>::parse(Cursor<Token> &) {
     return Ok(true);
 }
 
-void ValueParser<i32>::usage(Io::Emit &e) {
+void ValueParser<isize>::usage(Io::Emit &e) {
     e("integer");
 }
 
-Res<i32> ValueParser<i32>::parse(Cursor<Token> &c) {
+Res<isize> ValueParser<isize>::parse(Cursor<Token> &c) {
     if (c.ended() or c->kind != Token::OPERAND)
         return Error::other("missing value");
